Created Factories and Level Hierarchy
Ships, projectiles, and enemies are created by the factory model. An
abstract class Level was derived from most of the code in the
GalaxyHunterView, and new Levels subclass this class and implement it's
abstract methods. Enemies are in the beginning stages of AI and
placement.

Ship responds to movement in the same way and fires it's weapon when the
spacebar is pressed.