Rectangular Collision Detection Started
Collision detection works with rectangles where the rectangle's x and  y
coordinate are in the upper left hand corner of the bitmap that
represents an object. A new class Entity was created as the superclass
of all objects in the game, and has a method to compare if one entity's
rectangle overlaps another entity's rectangle.

The game checks now if enemies are hit by projectiles and if the ship is
hit by an enemy. If the ship is hit by an enemy the game stops all
animation.